wahid

+++ pelanggan.php
@@ -70,7 +70,7 @@
                                             ?>
                                         </span></a>
                                     <div class="dropdown-menu shadow dropdown-menu-right animated--grow-in" role="menu">
-                                        <div class="dropdown-divider"></div><a class="dropdown-item" role="presentation" href="logout.php"><i class="fas fa-sign-out-alt fa-sm fa-fw mr-2 text-gray-400"></i>&nbsp;Logout</a>
+                                        <div class="dropdown-divider"></div><a class="dropdown-item" role="presentation" href="logout.php" onclick="return confirm('Anda yakin mau logout ?')" ><i class="fas fa-sign-out-alt fa-sm fa-fw mr-2 text-gray-400"></i>&nbsp;Logout</a>
                                     </div>
                                 </div>
                             </li>
@@ -116,7 +116,7 @@
                                                             <button class="btn btn-primary btn-success" type="button">
                                                                 <i class="fa fa-edit"></i>Edit</button>
                                                         </a>
-                                                        <a href='delete_pelanggan.php?id=<?php echo $row['no_ktp']; ?>'>
+                                                        <a href='delete_pelanggan.php?id=<?php echo $row['no_ktp']; ?>' onclick="return confirm('Anda yakin mau menghapus data pelanggan ini ?')">
                                                             <button class="btn btn-primary btn-danger" type="button">
                                                                 <i class="fa fa-trash"></i>Delete</button>
                                                         </a>
